apps/ai-news/pyproject.toml
[project]
name = "ai-news"
version = "0.1.0"
description = ""
authors = [
    {name = "go-fireball",email = "venkatachalapatee@outlook.com"}
]
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "moviepy (>=2.2.1,<3.0.0)",
    "openai (>=2.8.1,<3.0.0)",
    "python-dotenv (>=1.2.1,<2.0.0)"
]

[tool.poetry]
packages = [{include = "ai_news", from = "src"}]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
